Update project dependencies.
The node-sass dependency was updated to ensure support with the newer
LTS versions of NodeJS (10.x). The ncp and bestzip dependencies were
removed because of major security vulnerabilities, a lack of community
support for the libraries, and because webpack has plugins supporting
the same functionality.

Testing Done:
Using node 10, cleaned node_modules, and ran `npm ci` and `npm run`
build commands. Verified successful build and content of zip file.
Uploaded produced zip file to a vCD installation and verified successful
load and execution of plugin.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>