Allow cross-origin calls to a VCD instance.
A proxy configuration has been added for the /api and /cloudapi
endpoints.

Usage: Change the target properties in the src/proxy.conf.json file to
point to a valid VCD instance. To avoid tracking changes to this file,
the index can be modified:
> git update-index --assume-unchanged src/proxy.conf.json

Testing Done: Ran `ng serve` with unconfigured file. Verified that
server successfully started up, that a call to
http://localhost:4200/api/versions returned a 404 and that the proxy
mapping was logged in the console.

Then configured the targets to have a valid VCD enpoint, restarted the
server, and verified that a call to http://localhost:4200/api/versions
correctly returned the API versions supported by the target instance.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>