Add VCD credential constructs to container.
An implementation is provided for specifying either a set of user
credentials (user@tenant, password) for creating a session, or a JWT
token for using an existing session.

These credentials will be used at bootstrapping time to create the
proper auth token for injecting into the VcdApiClient instance of
plugins.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>