Enable the bootstrapping of plugins by this container.
This change is the actual magic. Or ninja-level engineering. Take your
pick. A config service has been introduced and is being used as an
APP_INITIALIZER to ensure that this entire process completes before the
application can complete its own bootstrapping.

The container bootstrapping that we care about consists of three steps
(currently):
* load any VCD credentials if available and establish an
AuthTokenHolderService instance for plugins
* load the actual plugin modules
* update the container's router configuration to provide access to new
modules

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>