[VACE-2722] Fix Customize Portal Plugin validation
Remove  the plugin file validation, because with the  incremental
loading in place the plugins will have the ability to upload any kind
of files. Also the server doens't have  validation on its side, which means
in case there are any security vulnerabilities the UI is not capable to
stop them.

Update @vcd/sdk to v0.12.0

Testing Done:
- Verfiy the plugin works e2e

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>