[SDK] When fail to load plugin using @vcd/ui-components
Issue
The existing reg exprestion in our plugin builder is
targeting all angular depencies when we talk making
them "external", in webpack context. This works fine
as long as the plugin depends on angular depencies
that are provided by the Core UI.
However is some cases some other missing angular libs might be used by the plugin, for example @angular/testing.

Fix
Specify list of dependencies in ticketing project,
so @angular/testing won't be a provided by the Core UI.
Note this dependency won't be executed in any way in
runtime.

Testing Done:
Build ticketing example and deploy it to vcd
Verify the plugin is loaded
Verify the plugin works

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>